+++ src/main/cov.c
@@ -98,7 +98,7 @@
 			    ysd /= n1;					\
 			    sum /= n1;					\
 			}						\
-			sum /= (sqrt(xsd) * sqrt(ysd));			\
+			sum /= (sqrtl(xsd) * sqrtl(ysd));	       	\
 			if(sum > 1.) sum = 1.;				\
 		    }							\
 		}							\
@@ -191,7 +191,7 @@ static void cov_pairwise2(int n, int ncx, int ncy, double *x, double *y,
 		    sum += (xx[k] - tmp);	\
 	     tmp = tmp + sum / nobs;		\
 	}					\
-	_X_##m [i] = tmp;			\
+	_X_##m [i] = (double)tmp;		\
     }
 
 /* This uses two passes for better accuracy */
@@ -213,7 +213,7 @@ static void cov_pairwise2(int n, int ncx, int ncy, double *x, double *y,
 		tmp = tmp + sum / n;		\
 	    }					\
 	}					\
-	_X_##m [i] = tmp;			\
+	_X_##m [i] = (double)tmp;		\
     }
 
 
@@ -240,7 +240,7 @@ cov_complete1(int n, int ncx, double *x, double *xm,
 		for (k = 0 ; k < n ; k++)
 		    if (ind[k] != 0)
 			sum += (xx[k] - xxm) * (yy[k] - yym);
-		ANS(j,i) = ANS(i,j) = sum / n1;
+		ANS(j,i) = ANS(i,j) = (double)(sum / n1);
 	    }
 	}
 	else { /* Kendall's tau */
@@ -253,7 +253,7 @@ cov_complete1(int n, int ncx, double *x, double *xm,
 			    if (ind[n1] != 0)
 				sum += sign(xx[k] - xx[n1])
 				     * sign(yy[k] - yy[n1]);
-		ANS(j,i) = ANS(i,j) = sum;
+		ANS(j,i) = ANS(i,j) = (double)sum;
 	    }
 	}
     }
@@ -270,7 +270,7 @@ cov_complete1(int n, int ncx, double *x, double *xm,
 		else {
 		    sum = ANS(i,j) / (xm[i] * xm[j]);
 		    if(sum > 1.) sum = 1.;
-		    ANS(j,i) = ANS(i,j) = sum;
+		    ANS(j,i) = ANS(i,j) = (double)sum;
 		}
 	    }
 	    ANS(i,i) = 1.0;
@@ -309,7 +309,7 @@ cov_na_1(int n, int ncx, double *x, double *xm,
 			sum = 0.;
 			for (k = 0 ; k < n ; k++)
 			    sum += (xx[k] - xxm) * (yy[k] - yym);
-			ANS(j,i) = ANS(i,j) = sum / n1;
+			ANS(j,i) = ANS(i,j) = (double)(sum / n1);
 		    }
 	    }
 	    else { /* Kendall's tau */
@@ -322,7 +322,7 @@ cov_na_1(int n, int ncx, double *x, double *xm,
 			for (k = 0 ; k < n ; k++)
 			    for (n1 = 0 ; n1 < n ; n1++)
 				sum += sign(xx[k] - xx[n1]) * sign(yy[k] - yy[n1]);
-			ANS(j,i) = ANS(i,j) = sum;
+			ANS(j,i) = ANS(i,j) = (double)sum;
 		    }
 	    }
 	}
@@ -340,7 +340,7 @@ cov_na_1(int n, int ncx, double *x, double *xm,
 		else {
 		    sum = ANS(i,j) / (xm[i] * xm[j]);
 		    if(sum > 1.) sum = 1.;
-		    ANS(j,i) = ANS(i,j) = sum;
+		    ANS(j,i) = ANS(i,j) = (double)sum;
 		}
 	    }
 	    ANS(i,i) = 1.0;
@@ -371,7 +371,7 @@ cov_complete2(int n, int ncx, int ncy, double *x, double *y,
 		for (k = 0 ; k < n ; k++)
 		    if (ind[k] != 0)
 			sum += (xx[k] - xxm) * (yy[k] - yym);
-		ANS(i,j) = sum / n1;
+		ANS(i,j) = (double)(sum / n1);
 	    }
 	}
 	else { /* Kendall's tau */
@@ -384,7 +384,7 @@ cov_complete2(int n, int ncx, int ncy, double *x, double *y,
 			    if (ind[n1] != 0)
 				sum += sign(xx[k] - xx[n1])
  				    * sign(yy[k] - yy[n1]);
-		ANS(i,j) = sum;
+		ANS(i,j) = (double)sum;
 	    }
 	}
     }
@@ -409,7 +409,7 @@ cov_complete2(int n, int ncx, int ncy, double *x, double *y,
 			    if (ind[n1] != 0 &&	 xx[k] != xx[n1])	\
 				sum ++; /* = sign(. - .)^2 */		\
 	    }								\
-	    _X_##m [i] = sqrt(sum);					\
+	    _X_##m [i] = (double)sqrtl(sum);				\
 	}
 
 	COV_SDEV(x); /* -> xm[.] */
@@ -460,7 +460,7 @@ cov_na_2(int n, int ncx, int ncy, double *x, double *y,
 			sum = 0.;
 			for (k = 0 ; k < n ; k++)
 			    sum += (xx[k] - xxm) * (yy[k] - yym);
-			ANS(i,j) = sum / n1;
+			ANS(i,j) = (double)(sum / n1);
 		    }
 	    }
 	    else { /* Kendall's tau */
@@ -473,7 +473,7 @@ cov_na_2(int n, int ncx, int ncy, double *x, double *y,
 			for (k = 0 ; k < n ; k++)
 			    for (n1 = 0 ; n1 < n ; n1++)
 				sum += sign(xx[k] - xx[n1]) * sign(yy[k] - yy[n1]);
-			ANS(i,j) = sum;
+			ANS(i,j) = (double)sum;
 		    }
 	    }
 	}
@@ -498,7 +498,7 @@ cov_na_2(int n, int ncx, int ncy, double *x, double *y,
 			    if (xx[k] != xx[n1])			\
 				sum ++; /* = sign(. - .)^2 */		\
 		}							\
-		_X_##m [i] = sqrt(sum);					\
+		_X_##m [i] = (double) sqrtl(sum);			\
 	    }
 
 	COV_SDEV(x); /* -> xm[.] */
