+++ src/main/iosupport.c
@@ -192,7 +192,7 @@ int attribute_hidden R_TextBufferInit(TextBuffer *txtb, SEXP text)
 	l = 0;
 	for (i = 0; i < n; i++) {
 	    if (STRING_ELT(text, i) != R_NilValue) {
-		k = strlen(translateChar(STRING_ELT(text, i)));
+		k = (int) strlen(translateChar(STRING_ELT(text, i)));
 		if (k > l)
 		    l = k;
 	    }
